scripts/training/requirements_tf24odapi.txt
absl-py==0.10.0
apache-beam==2.27.0
astunparse==1.6.3
attrs==20.3.0
avro-python3==1.9.2.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
dill==0.3.1.1
dm-tree==0.1.5
docopt==0.6.2
fastavro==1.3.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.7.0
google-cloud-core==1.5.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
importlib-metadata==3.4.0
importlib-resources==5.1.0
joblib==1.0.0
kaggle==1.5.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.4
mock==2.0.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
object-detection @ file:///home/wendt/tf2odapi/models/research
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
pbr==5.5.1
Pillow==8.1.0
promise==2.3
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.8.0
py-cpuinfo==7.0.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.1
scipy==1.5.4
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
tensorflow-metadata==0.27.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tqdm==4.56.0
typeguard==2.10.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
